Projects/Flight/requirements.txt
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pandas==2.1.0
